Clean up /cart API routes
Reset the cart API controller methods to omit payment and order
processing logic for now. Restarting with RESTful CRUD API logic for the
shopping cart, which uses the new Redis session storage.
[Finishes #119419635]
